Rename the main state to "level"
