+++ index.js
@@ -5,9 +5,9 @@
 const config = require('./config');
 
 const endState           = require('./states/end');
+const levelState         = require('./states/level');
 const levelFailMenuState = require('./states/level-fail-menu');
 const loadState          = require('./states/load');
-const mainState          = require('./states/main');
 const mainMenuState      = require('./states/main-menu');
 const sceneState         = require('./states/scene');
 
@@ -22,7 +22,7 @@ document.addEventListener('DOMContentLoaded', function startGame() {
 
   game.state.add('scene', sceneState);
 
-  game.state.add('main', mainState);
+  game.state.add('level', levelState);
 
   game.state.add('main-menu',       mainMenuState);
   game.state.add('level-fail-menu', levelFailMenuState);
